+++ citests/citests/tests/__init__.py
@@ -1,6 +1,6 @@
 from .adapter_tests import DiscChiSq, DiscGSq, DummyFisherZ, HarteminkChiSq
 from .contingency_table_tests import ChiSq, GSq
-from .kernel_tests import KCI, RCIT, RCoT
+from .kernel_tests import GKCM, KCI, RCIT, RCoT
 from .ml_based_tests import GCM, PCM, WGCM
 from .nearest_neighbor_tests import CMIknn, CMIknnMixed, MCMIknn
 from .partial_correlation_tests import FisherZ, Spearman
@@ -11,7 +11,7 @@ __all__ = [
     "ChiSq", "GSq",
     "RegressionCI", "CiMM",
     "CMIknn", "CMIknnMixed", "MCMIknn",
-    "KCI", "RCIT", "RCoT",
+    "KCI", "RCIT", "RCoT", "GKCM",
     "GCM", "WGCM", "PCM",
     "DiscChiSq", "DiscGSq", "DummyFisherZ", "HarteminkChiSq",
 ]